lsp: Rename SubprocessSphinxClient to SphinxClient
Originally the plan was to have multiple implementations of the
SphinxClient interface, one where Sphinx was run in-process and
another where it in a separate subprocess.

Only one was ever implemented and the layer of indirection is just a
hinderance at this point.
